route4: use rib to list routes
A small quirk is required to properly list default routes.
rte_rib_get_nxt does not return a default route since it will only
return routes that have larger prefixes/depths than the one passed as
parameter.

This probably warrants a fix in rte_rib but it is not trivial.

In the meantime, add an explicit rte_rib_lookup_exact() for 0.0.0.0/0
and include it if it returns a route entry.

Link: http://inbox.dpdk.org/dev/CZY0LHKYV4ZO.YRBZ6C6GW6AI@redhat.com/
Signed-off-by: Robin Jarry <rjarry@redhat.com>