api: remove client library
Cleanup header names. Move them all to a suitable "api" folder.

Remove the libbr.so client library. It serves very little purpose. Move
most of its code directly in the cli functions.

For common code that cannot be part of headers, move it to
a br_api_client_impl.h header that must be included in the client
application code so that it defines the br_api_client_* helper
functions.

Signed-off-by: Robin Jarry <rjarry@redhat.com>