ip4: add basic arp support
It turns out that supporting ARP drags a lot of stuff.

A proper way to declare local IP addresses
==========================================

Without a local address, we cannot reply to ARP requests. Also, a local
address is required to *send* ARP requests.

Numerous extra fields in next hop objects
=========================================

The chosen implementation is to reuse next hops to store port addresses
with a special flag (BR_IP4_NH_F_LINK) along with their network prefix
length.

A "connected" route is added for every address and its next hop is the
local address. A pointer to the next hop is also stored in an array,
indexed by port_id.

Reworking how IP routing is done
================================

A route lookup is performed for all packets in ip_input. If we have
a valid next hop for the destination IP, the packet is either passed to
ip_forward or ip_local (depending on the BR_IP4_NH_F_LOCAL flag and the
IP address of the next hop).

ip_forward now only decrements TTL and recomputes the IP checksum.

ip_output will now reuse the resolved next hop from ip_input.
If the next hop is a BR_IP4_NH_F_LINK, a new next hop is created for the
destination IP along with its associated /32 route. This way, next time
we need to forward/output such a packet, we'll get the correct next hop
in a single route lookup.

If the next hop is reachable, the packet will be passed to eth_tx after
updating the eth private data.

If it is not reachable, the packet is passed to arp_output_request. If
possible, the packet is appended to a hold queue associated with the
next hop. The hold queue is implemented as a linked list limited in
length.

A new ARP request packet is then forged and passed to eth_tx.

Handling of ARP packets in data plane
=====================================

ARP ether type is registered and associated with the arp_input node.
This node deals with updating the next hops identified by the ARP packet
source IP. Both ARP requests and replies are considered when updating
next hops. If a next hop does not exist and the target IP is local
(assigned to a port), a new next hop and its associated /32 route are
created dynamically.

When a next hop is updated, its queue of held packets is flushed and
sent to ip_output. These packets will be processed as before (route
lookup), only with their next hop now being valid.

If the received ARP packet is a request, it is passed to
arp_output_reply which deals with formatting a reply based on the
requested ARP target IP address.

Not implemented yet
===================

* Aging of dynamically learned ARP entries (and their /32 routes).
* Automatic refresh of next hops before they expire.
* Retry of ARP requests in the background after a grace period.
* Filtering of the (hidden?) dynamic /32 next hops & routes.

Signed-off-by: Robin Jarry <rjarry@redhat.com>