arp: avoid race conditions in hold queue
If we receive an ARP reply in another thread while sending an ARP
request we may queue a packet while we could have sent it. Check for
BR_IP4_NH_F_REACHABLE *and* enqueue the packet while holding the next
hop lock. If the next hop somehow became reachable, send the packet
instead of queuing it.

Signed-off-by: Robin Jarry <rjarry@redhat.com>