ipip: fix tunnel hash key structure layout
Using uint16_t causes the compiler to add 2 bytes padding at the end of
the structure. When the structure is initialized on the stack, the
padding bytes have undetermined contents depending on what is present on
the stack.

When inserting the tunnel key on interface creation, these two padding
bytes may contain different data than when calling ipip_get_iface in the
data path. This causes confusing and seemingly random IPIP tunnel lookup
failures.

This structure is used to compute a hash key. In order to get
deterministic results, use uint32_t to store the vrf_id so that the
compiler does not insert any padding.

Signed-off-by: Robin Jarry <rjarry@redhat.com>