nexthop: fix nexthop hash key structure layout
Using uint16_t for vrf_id causes the compiler to add 2 bytes padding in
the middle of the structure. When the structure is initialized on the
stack, the padding bytes have undetermined contents depending on what is
present on the stack.

When calling ip4_nexthop_add, these two padding bytes may contain
different data than when calling ip4_nexthop_lookup.

This structure is used to compute a hash key. In order to get
deterministic results, use uint32_t to store the vrf_id so that the
compiler does not insert any padding. Put the field at the end to be
consistent with ipip.

Signed-off-by: Robin Jarry <rjarry@redhat.com>