iface: fix structure sizes and alignment
Fix misaligned accesses as reported by libubsan:

 modules/infra/control/port.c:439:22: runtime error: member access
 within misaligned address 0x00017f63dccc for type 'const struct
 iface_info_port', which requires 8 byte alignment

Force the alignment of the iface.info field to 8. Also force the
alignment of all iface_info_* structures to 8 as well (it was already
like this, but at least now, it is explicit).

To fit the iface structure into a single cache line, replace the name
and devargs arrays to stack allocated strings and only store the
pointers. This saves (64*2)-(8*2) == 112 bytes per iface+port.

Signed-off-by: Robin Jarry <rjarry@redhat.com>