infra: add wrapper to create mempools
Instead of creating dedicated mempools for each port and feature,
use a wrapper to allocate a reduced number of mempools.

This will avoid corner cases where the mempool for control input (eg
512) is depleted because the mbuf aren't released by the tx_burst function
immediatly.
These elements are sent back to the mempool only when a threshold is reached,
and this depends on the queue size.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>