iface: use dynamic array for subinterfaces
Using a simple counter makes it complex to track individual interfaces.
In case of error, we may increment/decrement the subinterface counter
multiple times for the same subinterface.

Use a pointer array instead of a counter. Only add a subinterface if it
isn't already present and only remove it if it is present.

Signed-off-by: Robin Jarry <robin@jarry.cc>