ip: fix source address for forward errors
The source addressed used to forge an ICMP packet should be based on
a route lookup for the original packet source address. This provides
a next hop which has an outgoing interface and an IP address.

With that IP address, we can determine what is the best interface
address to use as source.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>
Acked-by: Robin Jarry <robin@jarry.cc>