ip: ensure input_iface is not set to garbage
If the route lookup fails in ip_input, an ICMP destination unreachable
message is sent to the source address. To do this, we need the input
interface to determine the correct interface address with which to reply
as source address.

Make sure that input_iface is set to the correct value or force it to
NULL if we do not have the information (arp_input, icmp_output).
The input_iface value in ipip_output needs to be adjusted as well.

Signed-off-by: Robin Jarry <rjarry@redhat.com>