ip: reject unicast packets not addressed to local mac
Grout is not a L2 switch. It should never forward packets that have not
been sent to its receiving input interfaces. Inspect the destination
ethernet address in eth_input and set a new eth_dst field in the
eth_input_mbuf_data structure according to the address.

ETH_DST_LOCAL // destination is the input interface mac
ETH_DST_BROADCAST // destination is ff:ff:ff:ff:ff:ff
ETH_DST_MULTICAST // destination is a multicast ethernet address
ETH_DST_OTHER // destination is *not* the input interface mac

In ip_input, check the value of eth_dst and immediately send to
ip_input_local in the case of broadcast or multicast. Drop the packet if
it is set to ETH_DST_OTHER.

There is a special case for ipip_input which needs to also set eth_dst
to ETH_DST_LOCAL so that the packet is not dropped in ip_input.

Signed-off-by: Robin Jarry <rjarry@redhat.com>