ci: add missing dependencies for mlx5 and ice
DPDK needs to be linked with libarchive in order to unpack the ICE
firmware.

  EAL: libarchive not linked, /lib/firmware/intel/ice/ddp/ice.pkg.xz
  cannot be decompressed

When enable_drivers contains wildcards, the explicitly enabled drivers
are silently disabled if some of their system dependencies are missing.
The mlx5 driver requires libibverbs.

Add missing dependencies in README and in the CI jobs. Remove wildcards
from enable_drivers to ensure the build fails when system dependencies
are missing.

Reported-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: Robin Jarry <rjarry@redhat.com>