main: fix syslog redirection when running in a graphical session
When running in a graphical session managed by a systemd slice, the
JOURNAL_STREAM environment variable will be set and grout will
wrongfully assume it should redirect its logging to syslog.

Only redirect logging to syslog when INVOCATION_ID is set which means
that the process was started as a systemd service.

Fixes: 596fe8637360 ("main: restrict logging to syslog when running under systemd")
Signed-off-by: Robin Jarry <rjarry@redhat.com>