main: allow creating ports when only a single cpu is available
When starting grout with a single CPU in the thread affinity, the main
lcore uses it and no port can be created since worker_ensure_default()
returns a ERANGE error (no additional CPUs available).

When creating a port and no CPU can be found that is in the correct
socket and not equal to main_lcore. Fallback on whatever is available
and display a warning in the logs. The port rxqs can then be moved to
isolated CPUs manually.

Also, rte_eal_init() changes the thread affinity of the first lcore.
Since this main lcore is not used for datapath, only for control plane
operations, it is better to let the kernel do the scheduling on other
CPUs if possible. Remember the original startup thread affinity before
calling rte_eal_init() and restore it on the main thread.

Signed-off-by: Robin Jarry <rjarry@redhat.com>