ip6: fix show route
rte_rib6_get_nxt() needs an ipv6 prefix in the form of a pointer to
a masked rte_ipv6_addr. 0 is seen as NULL and both rte_rib6_lookup_exact
and rte_rib6_get_nxt return an error which is silently ignored.

Pass a proper zero address along with a 0 prefix length which means
"all" the IPv6 addressing space.

Fixes: 0fc1cd222b8e ("modules: add static ipv6 forwarding support")
Signed-off-by: Robin Jarry <rjarry@redhat.com>