infra: expose idle duration in stats
Add an extra fake "idle" node statistics holding the number of usleep()
calls that were performed along with the total number of cycles spent
sleeping.

The number of packets will always remain at 0 for this stat. The only
way to show it with grcli is to ask for zero stats:

  grout# show stats software zero pattern idle
  NODE    CALLS  PACKETS  PKTS/CALL  CYCLES/CALL  CYCLES/PKT
  idle  2403813        0        0.0     101462.3         0.0

Signed-off-by: Robin Jarry <rjarry@redhat.com>