datapath: increase sleep threshold to 256 empty polls
Checking for "empty polls" every 32 rounds of main loop is too frequent.
rte_graph_walk() can be very fast when there are no packets and this
leads to usleep(1) being called while traffic is present (but only
absent for 32 loops).

To ensure we only go to sleep less often, increase the required number
of "empty polls" up to 256.

Signed-off-by: Robin Jarry <rjarry@redhat.com>