treewide: homogenize how unused parameters are identified
Remove all occurrences of:

	(void)unused_param;

Since these can be confusing and do not convey any meaning about the
variable being unused. It is a legacy way to silence the
unused-parameter compiler warning.

The whole grout code base is compiled with --std=gnu2x which allows
unnamed parameters even in function definitions. Make use of that to
indicate unused parameters. When the parameter type is explicit (e.g.
struct rte_graph *), simply remove the parameter name which is usually
redundant, e.g.:

	void foo(const struct rte_graph *, struct rte_node *node) { }

When the type is not explicit (e.g. int), keep the parameter
wrapped in a comment, e.g.:

	int foo(void * /*cookie*/, const char *buf, size_t size) { }

Update CONTRIBUTING.md for posterity.

Reported-by: Aaron Conole <aconole@redhat.com>
Signed-off-by: Robin Jarry <rjarry@redhat.com>
Acked-by: David Marchand <david.marchand@redhat.com>