devtools: fix check-patches with dash
If dash is used as the default shell interpreter at /bin/sh, which is
more POSIX-compliant than bash, the hook related to the issue check
might fail if one field of the JSON output (like "body") contains
multiples lines.

Example with the issue 78:

  jq: parse error: Invalid string: control characters from U+0000
      through U+001F must be escaped at line 130, column 4
  error [PATCH 1/1]  [...] 'https://github.com/DPDK/grout/issues/78'
        does not reference a valid open issue

From POSIX spec for echo:

  It is not possible to use echo portably across all POSIX systems
  unless both -n (as the first argument) and escape sequences are
  omitted.
  [...]
  New applications are encouraged to use printf instead of echo.

Use printf instead of echo to be more portable.

Signed-off-by: Julien Meunier <dev@julienmeunier.eu>
Acked-by: Robin Jarry <robin@jarry.cc>