iface: avoid use after free on duplicate name
When an interface name is already taken, we return an error but the
iface pointer is set the already existing interface. Before returning
from iface_create, that pointer is freed without removing the interface
from the system.

When listing interfaces or accessing anything related to interfaces
after that causes an invalid memory access.

Reset iface to NULL before returning an error to avoid that.

Fixes: a0f9d6699605 ("iface: forbid duplicate names on creation")
Reported-by: Christophe Fontaine <cfontain@redhat.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>