infra: implement packet tracing for base nodes
Implement packet tracing for the infrastructure nodes: rx, tx, drop,
control_input and control_output.

Check if the GR_IFACE_F_PACKET_TRACE flag is set on the receiving
interface in the rx node to determine if tracing should be started for
each mbuf.

For control_input, rely on gr_trace_all_enabled() to decide if we enable
tracing.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Co-authored-by: Christophe Fontaine <cfontain@redhat.com>
Acked-by: David Marchand <david.marchand@redhat.com>