icmp6: implement packet tracing
Store the full ICMPv6 header as tracing data. The header is of variable
size depending on the ICMPv6 type. Use the IPv6 payload length field
(after accounting of optional extension headers) to determine what is
the correct amount of data to store as tracing data.

Reuse the trace_icmp6_format function introduced in the previous commit
as the trace_format callback.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: David Marchand <david.marchand@redhat.com>