trace: make writing to buffer safer
Define a SAFE_BUF macro to allow calling snprintf-like functions and
catch potential errors without making the code an unreadable mess of if
conditions and goto statements.

It is very opinionated about the variable and goto label names but it
should be fine for most uses. The only parameters besides the variable
arguments are the function name and the buffer size.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: David Marchand <david.marchand@redhat.com>