main: implement basic dynamic array utilities
We currently depend on stb_ds.h for certain non-critical code paths.
This lib is not available in any distribution and basically would
require grout to vendor it. Since it provides a lot more features that
we actually need. Implement a stripped down version of it. Do not try to
implement any of the hash map features. We do not need them.

Add basic unit tests.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Christophe Fontaine <cfontain@redhat.com>