arp: fix nexthop lookup
When we have a prefix route associated with a local interface, route
lookup may return a local nexthop.  Returned nexthop IP doesn't
correspond to the arp request/reply ip address.  This causes the ARP
packet to be dropped without updating any next hop.

Do not look for a route, but for a nexthop matching the IP address.

Fixes: bfae0e36a7b4 ("ip: fix unresolved arp management")
Closes: https://github.com/DPDK/grout/issues/89
Signed-off-by: Christophe Fontaine <cfontain@redhat.com>
Acked-by: Robin Jarry <robin@jarry.cc>