meson: make scdoc optional
Add a meson feature option to control if man pages are generated using
scdoc. By default, the feature is set to "auto". In this case, if scdoc
is not installed, the man pages will not be built.

If set to "enabled", scdoc will be required as it was before this patch.

Add --auto-features=enabled to the debian package which sets all "auto"
features to "enabled" to require the dependency as before. RPM meson
macros already do this.

Signed-off-by: Vincent Jardin <vjardin@free.fr>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: David Marchand <david.marchand@redhat.com>