meson: make cmocka optional
Add a meson feature option to control if unit tests are build and linked
to cmocka. By default, the feature is set to "auto". In this case, if
cmocka is not installed, the unit tests will not be built.

If set to "enabled", cmocka will be required as it was before this
patch.

Signed-off-by: Vincent Jardin <vjardin@free.fr>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: David Marchand <david.marchand@redhat.com>