smoke: check that docs/graph.svg is up to date
Dumping the graph requires that grout is running and has at least one
interface created. Smoke tests seems like a good place do do so. Add
a minimal test that just creates an null port and dumps the graph.

Pre-process the svg output to extract all edges. Compare the runtime
output from grcli to the one stored in git. Fail the test if differences
are found.

Update README.md and CONTRIBUTING.md with instructions to update the
reference graph image.

Signed-off-by: Robin Jarry <robin@jarry.cc>