deb,rpm: fix build error when no tag is found
When no tags are found in the current repository (or when not building
from a git repo), git describe --long --abbrev=8 --dirty will return an
error and the obscure sed command will be executed instead to determine
the current version.

When hostname is not installed, it leads to an even more obscure error:

 /bin/sh: line 1: hostname: command not found
 rpmbuild -bb --build-in-place -D 'version 0.2' -D 'release 20241202102621..el9' rpm/grout.spec
 error: line 14: Illegal sequence ".." in: Release: 20241202102621..el9
 make: *** [GNUmakefile:72: rpm] Error 1

Replace the hostname value with a hard-coded "local" string. The machine
where this build took place does not matter.

Fixes: 763725c614cc ("deb,rpm: fix error when no tag is found in local repo")
Signed-off-by: Robin Jarry <robin@jarry.cc>