address: use common hoplist structure
Now that both ipv4 and ipv6 use the same nexthop structure, we can
replace the hoplist6 structure with hoplist. Move hoplist definition
into the infra module and update ipv6 interface address management
accordingly.

Add a note next to hoplist that it could be improved with a dynamic
vector instead of a fixed size array.

Signed-off-by: Robin Jarry <robin@jarry.cc>