infra: extend control input to inject rte_mbufs
With "control_input", the control plane is able to inject a pointer to
the dataplane ; in order to meet the expectation that all nodes will
receive a rte_mbuf* array, we allocate the mbuf in the source node
"control_input_process".

Yet, there are some cases where we want to allocate and fill a rte_mbuf
from the control plane thread, to avoid a copy later in the dataplane
thread.
In order to differentiate if the node expects an abstract pointer or a
mbuf, modify the register function to accomodate both requirements,
Update all users of "gr_control_input_register_handler".

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>