nexthop: fix weird blackhole constant value
The BLACKHOLE=(IP4_MAX_NEXT_HOPS + 1) constant dates back prior to when
we started storing fib nexthop values as pointers instead of offsets in
a fixed size array.

Having a nexthop pointer being an array offset does not make any sense.
Replace it with zero to be consistent with NULL pointers.

Signed-off-by: Robin Jarry <robin@jarry.cc>