ip6/output: handle unresolved nexthops in control plane
Do not modify nexthop objects in the data plane. It requires acquiring
locks and also modifying the routing table. It can lead to races between
multiple worker threads.

Instead, when a nexthop is unreachable in ip_output, send the mbuf to
the control plane thread and do whatever is required there:

* Possibly create a new nexthop for prefix routes and create a new /128
  route that points to it.
* Hold the packet to the nexthop waiting for NDP resolution.
* Send an NDP request probe (if one hasn't already been sent earlier).

Change all nexthop references to const in the datapath to ensure they
cannot be modified.

Signed-off-by: Robin Jarry <robin@jarry.cc>