arp: reply to all local iface addresses
Do not limit replying to ARP requests to addresses configured on the
ingress interface. If the target address is a local IP addresses
configured on any interface, send a reply on the ingress interface.

Update arp_output_reply to allow sending a reply for an address which is
not on the ingress interface. arp_mbuf_data only needs one local nexthop
reference.

Link: https://datatracker.ietf.org/doc/html/rfc826
Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Christophe Fontaine <cfontain@redhat.com>