ip6: answer to router sollicit message
Upon reception of a router sollicit message, trigger the existing
timer to send immediately a router advertismement packet.

Add relevant test: as the same mac address is defined for 2 interfaces
in the 2 namespaces, the same link local ip address will be generated
and used by linux.

Update github workflow and doc to reflect the addition of 'ndisc6'
as a test tool.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>