ip6: scoped route and nexthop lookup
We may have a similar linklocal next hop (same ip, same mac) on 2
different interfaces.
As the same IP can't exist twice in the FIB, the nexthop is updated
and goes from 1 interface to the other.

Fix this by scoping the address, allowing the insertion of the peer
address once per interface.
The IP in the nexthops are *not* scoped, as we already have the
interface in the structure. Yet, in order to differentiate 2 nexthops
with the same IP, add a new parameter to the nexthop_lookup and
route_lookup fonctions

Scope properly the target addresses and remove the interface id when
necessary before emitting the packets.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>