ip6: scope peer link local addresses
We may have a similar linklocal next hop (same ip, same mac) on 2
different interfaces.
As the same IP can't exist twice in the FIB, the nexthop is updated
and goes from 1 interface to the other.

Fix this by scoping the address, allowing the insertion of the peer
address once per interface.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>