loopback: remove ip/ipv6 calls from generic node
The loopback_input node is in the infra module and should not depend on
the ip nor ip6 modules.

Add a dynamic register mechanism like eth_input to allow the
ip{,6}_input nodes to plug themselves after loopback_input.

To prevent packets received on a loopback interface from being
forwarded, add a new ETH_DOMAIN_LOOPBACK flag to be set instead of
ETH_DOMAIN_LOCAL in the mbuf private data and act accordingly in
ip{,6}_input.

Signed-off-by: Robin Jarry <robin@jarry.cc>