ndp: properly set output iface for neigh solicits
When sending a neighbor solicitation from the control plane the
destination IPv6 address will be multicast.

Since the packet originates from the control plane, it does not have any
input interface set.

This causes ip6_output to ignore the nexthop->iface_id and rely only on
mbuf_data->iface which is NULL and eventually drops the packet in
ip6_error.

Set the interface of the solicited nexthop as "input" interface. It will
be used in ip6_output to properly determine which source mac address to
use.

Fixes: 07ff3ca67b29 ("ip6: fix iface output for multicast destination")
Signed-off-by: Robin Jarry <robin@jarry.cc>