event: allow internal subscriptions
Allow internal code to subscribe to events. Define a new event
subscription structure that holds a callback and a list of events to
subscribe to. The callback is invoked with an event type ID and an
opaque const void * value.

On registration, event subscriptions are stored using a linked list
which is iterated on event push to determine which callbacks to invoke.

The API socket event subscription becomes a subscriber to this new
system using the special EVENT_TYPE_ALL value to be called for all
events.

The API socket subscribers need the opaque const void * value to be
"serialized" to data suitable for consumption (i.e. structures defined
in gr_api.h and consorts). For this, event serializers must also be
registered per event type (similar to event subscribers). They must
define a callback that will convert an internal const void * value into
a malloc'd buffer. The buffer will be freed by the API socket event
callback after it has been sent to all subscribers.

For trivial conversions (IP and IPv6 routes), only report the size and
let generic code perform a memcpy.

Replace all event push functions and code accordingly.

Signed-off-by: Robin Jarry <robin@jarry.cc>