main: add flag to redirect to syslog
Relying on environment variables to determine whether logs should be
redirected to syslog/journald is unreliable.

For example, desktop environments may be started as systemd transient
services with systemd-cat and INVOCATION_ID may be present in the
environment variables without the user/developer having any control over
it.

Instead of trying to guess in which context grout is executed, add a new
--syslog flag that will explicitly redirect logs to syslog.

Signed-off-by: Robin Jarry <robin@jarry.cc>