ip6_local: fix extension header parsing
rte_pktmbuf_read() needs a buffer length and a pointer to the buffer.
_ext is an array of uint8_t of size 4. sizeof(&_ext) returns the size of
a pointer (8 bytes). This means that rte_pktmbuf_read() will potentially
overflow on the stack if the packet is not contiguous.

Fix the call to rte_pktmbuf_read() by passing sizeof(_ext) and _ext
without prefixing them with &.

Fixes: 520047d56e4c ("ip6: fix IPv6 extension header parsing")
Signed-off-by: Robin Jarry <robin@jarry.cc>