ip6_local: report an explicit error when reading packet fails
If reading from the mbuf fails, stop processing immediately and drop the
packet with an explicit ip6_input_local_error stat.

Do not adjust the mbuf offset in case of early error, it is not
necessary.

Move the trace add to the beginning of the loop to ensure it is always
called.

Signed-off-by: Robin Jarry <robin@jarry.cc>