infra: set iova mode
When running grout in a VM and hotplugging a port, the IOVA mode
used is always VA, which leads to the impossibility to bind the port.

ERR: PCI_BUS:   Expecting 'PA' IOVA mode but current mode is 'VA', not initializing

Check if 'noiommu' is enabled, and sets iovamode to "PA".

Suggested-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: Christophe Fontaine <cfontain@redhat.com>