c2x: use typed enum instead of compiler internals
__attribute__((mode(QI/HI/...))) allow to specify the size
of an enum.
Yet, enums are still *signed* values. mode(QI) implies implies
that the valid values are in the range [-127..128].
This creates an issue with clang for values above 128, such as icmp
codes.

With c2x standard, similar to C++, we can specify the backend of an
enum, with the proper "unsigned" attribute.
enum foo: unsigned char {a, b, c};

Replace all mode(QI/HI) with the proper type, so clang doesn't complain.

This commit increases minimal gcc version to gcc 13.

Signed-off-by: Christophe Fontaine <cfontain@redhat.com>