ip: enhance ip header validation
The current code assumes that the hardware performs the following checks:
- The packet length is sufficient to accommodate the minimum length of the
  IP header.
- The IP version is set to 4.
- The IP header length field is large enough to accommodate the minimum
  legal IP datagram length.

However, these assumptions are incorrect, at least with the tap driver. The
following packets can be created using scapy:

> p_badversion = Ether(dst='f0:0d:ac:dc:00:00', src='ba:d0:ca:ca:00:00')/IP(version=5, src='172.16.0.2', dst='172.16.1.2')
> p_badihl = Ether(dst='f0:0d:ac:dc:00:00', src='ba:d0:ca:ca:00:00')/IP(src='172.16.0.2', dst='172.16.1.2', ihl=3)
> p_truncated = Ether(dst='f0:0d:ac:dc:00:00', src='ba:d0:ca:ca:00:00')/IP(src='172.16.0.2', dst='172.16.1.2')[:14+10]

If the correct routes are configured, grout will forward these packets instead of discarding them,
in violation of RFC1812.

So software validation is required for these fields.

The next commit will use packet type to not checking check IP version and IHL.

Fixes: 5f5de2c05188 ("ip4_lookup: comply with rfc1812")
Signed-off-by: Maxime Leroy <maxime@leroys.fr>