tap: force equal number rxqs and txqs
The tap driver does not support a different number of RXQ and TXQ.

Force the same value in both (using the maximum between the two).

This is only a workaround. This means that when the number of queues
increases (either because the number of workers increases or because the
user explicitly sets a higher number of RQXs), there will not be any way
to reduce the number of queues.

There is unfortunately no way to know what is a good value. Since tap
ports are only used in tests, it should be OK for now.

Signed-off-by: Robin Jarry <robin@jarry.cc>