ip: add unit tests to validate ip header
This patch introduces unit tests to validate the correctness of the IPv4
header processing in the ip_process_input function. The tests check for
key properties such as IP version, checksum calculation, packet length,
and other essential fields.

To test ip_input_process, the `rte_node_enqueue_x1` function needs to be
wrapped. However, since it is defined as a `static inline` function in the
original code, it gets expanded at compile-time and does not have a
linkable symbol. This prevents direct wrapping using CMocka’s function
wrapping mechanism.

To work around this limitation, a `#define` is used to replace calls to
this function with a wrapped implementation inside the unit tests. This
ensures that all occurrences of `rte_node_enqueue_x1` in the tested code
are redirected to the wrapped version during unit testing.

Placing this #define in gr_graph.h for reuse in other unit tests seems
beneficial. However, it leads to compilation issues. The problem arises
because rte_node_enqueue_x1 definition gets expanded to
__wrap_rte_node_enqueue_x1 in rte_graph_worker_common.h, while
__wrap_rte_node_enqueue_x1 is already defined in gr_graph.h, causing a
redefinition error.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>