ip6: clarify usage of mpriv in ip_input_process
The eth_input_mbuf_data and ip6_output_input_mbuf_data pointers are
initialized at the same time, which may suggest that both structures can be
accessed concurrently. However, this is not the case, as
GR_MBUF_PRIV_DATA_TYPE defines the private type as a union.

To improve clarity, this patch ensures that ip6 variable is defined
only when this one is actually used.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>