ip6: demonstrate DPDK issue with IPv6 version field encoding
This patch is intended to highlight a problem when setting the version
field of the IPv6 header using DPDK. Specifically, it reveals unexpected
behavior potentially related to byte-order conversion or compiler
handling on certain architectures.

The patch is not intended for merging. It serves purely as a diagnostic
tool to reproduce the issue on ARM platforms and across different
compilers in the GitHub CI infrastructure.

This will help in reporting and reproducing the issue on DPDK Bugzilla.

Signed-off-by: Maxime Leroy <maxime@leroys.fr>