nexthop: fix ageing logic
The boolean logic for determining if a probe should be sent or not is
incorrect. We absolutely don't care if the nexthop is a gateway or not.
Moreover, request_age > probes does not make any sense. I may not have
been in full control of my brain at the time I wrote this.

Only try to send probe if a nexthop is either PENDING or STALE. If we
already sent the maximum number of probes, mark the nexthop as FAILED so
that it is candidate for deletion after one minute.

Fixes: 778bd652b3ed ("ip: implement arp aging and refresh")
Signed-off-by: Robin Jarry <robin@jarry.cc>